+++ src/Client.php
@@ -302,6 +302,13 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
     private function transfer(RequestInterface $request, array $options): PromiseInterface
     {
         $request = $this->applyOptions($request, $options);
+
+        if ('' === $request->getProtocolVersion()) {
+            trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Sending a request with an empty protocol version is deprecated; guzzlehttp/guzzle 8.0 will reject empty protocol versions.');
+
+            $request = Psr7\Utils::modifyRequest($request, ['version' => '1.1']);
+        }
+
         /** @var HandlerStack $handler */
         $handler = $options['handler'];
 
@@ -471,6 +478,12 @@ class Client implements ClientInterface, \Psr\Http\Client\ClientInterface
      */
     private static function normalizeProtocolVersion($version): string
     {
+        if ('' === $version) {
+            trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Passing an empty "version" request option is deprecated; guzzlehttp/guzzle 8.0 will reject empty protocol versions.');
+
+            return '1.1';
+        }
+
         return \is_float($version) ? \number_format($version, 1, '.', '') : (string) $version;
     }